requirements.txt
Django>=3.1.0
djangorestframework
requests